% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday}
\usage{
birthday(x)
}
\arguments{
\item{x}{a birthday or number out opf 365}
}
\value{
a probability from 0 to 1
}
\description{
Birthday
}
\examples{
birthday(20:25)
}
