% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT}
\format{
## `ddt`
A data frame with 144 rows and 6 columns:
\describe{
  \item{RIVER}{Name of river/water source sample was taken at}
  \item{MILE}{Distance in miles from river mouth}
  \item{SPECIES}{Species of fish}
  \item{LENGTH}{Fish length (cm)}
  \item{WEIGHT}{Fish weight (g)}
  \item{DDT}{Measured DDT concentration (ppm)}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
ddt
}
\description{
A dataset detailing fish, fish species, length they travel, river they travel
etcetera ... and ultimately their DDT concentration within the TN river valley
}
\keyword{datasets}
