% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIRE}
\format{
## `fire`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{DISTANCE}{Distance from fire station (miles)}
  \item{DAMAGE}{Damage in thousands of dollars}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
fire
}
\description{
A dataset detailing the fire damage and distance form the fire station
}
\keyword{datasets}
