% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotbinom.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of times the experiment is run}

\item{n}{number of options}

\item{p}{probability of a 1 or 0 occurring}
}
\value{
a barplot with the binomial values through the table
}
\description{
Title
}
\examples{
mybin(iter=1000,n=10,p=.5)
}
