% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybootadj.R
\name{mybootadj}
\alias{mybootadj}
\title{mybootadj}
\usage{
mybootadj(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations of the sample}

\item{x}{sample size}

\item{fun}{the function to be used (eg.: mean, variance)}

\item{alpha}{Confidence level (confidence = 1 - alpha)}

\item{cx}{a constant value}

\item{...}{Ability to include extra parameters}
}
\value{
A bootstrap function with a histogram that has confidence intervals and other sample statistics
}
\description{
mybootadj
}
\examples{
\dontrun{myboot2(iter=10000, x=sam, fun="mean", alpha=0.05, xlab="mean", col="#0f20db")}
}
