% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltunidistfunction.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Theorem Uniform Distribution Function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{amount of times the sample is taken and the math applied}

\item{a}{minimum value/lower limit of the distribution}

\item{b}{maximum value/upper limit of the distribution}
}
\value{
a histogram plot of the uniform central limit theorem
}
\description{
Central Limit Theorem Uniform Distribution Function
}
\examples{
\dontrun{mycltu(n=20, iter=10000)}
}
