% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{value to compute P(x <= a)}
}
\value{
a list with the mean, standard deviation, a, and area components
}
\description{
myncurve
}
\examples{
\dontrun{mycurve(mu=5, sigma=3, a=6)}
}
