% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Seats available on the flight}

\item{gamma}{Probability of overbooking}

\item{p}{Likelihood a passenger will show}
}
\value{
Plots for both discrete and continuous methods along with a list of variable values
}
\description{
Title
}
\examples{
\dontrun{ntickets(N=400, gamma = .02, p = .95)}
}
